Add Walking Animation
Added the walking animation for the main character and edited the code so that the animation only plays when the character is on the ground and is moving left or right. Also increased the linear drag so the character doesn't slide around and increased the jump height to compensate for the lost velocity.

Co-authored-by: day00090 <day00090@morris.umn.edu>