Modify game UI and fix health
Modified the game UI by formatting the second and minutes elapsed in the game with mm:ss. Additionally, fixed the health bar so it successfully removes health visually once the player gets hit. There is also a new basic square sprite to be used for the health bar so it maintains a square shape.

Co-authored-by: day00090 <day00090@morris.umn.edu>