Add new Slime Enemy
Adds a new slime enemy to the game. Also modifies the way we calculate the position when adding a new ally to the top of the stack. The new position calculator isn't working how we expect it to, so it is a work in progress and will likely be changed in the future.

Co-authored-by: ErikRauer <rauer007@morris.umn.edu>