Fix Ally Stacking
Update the PlayerController and EnemyController script so that the allies properly stack on the player, and the allies no longer overlap (too much). The bug with the allies stacking over empty space after losing an ally is fixed.
We also fixed the slime sprite so that the center is actually in the center.

Co-authored-by: ErikRauer <rauer007@morris.umn.edu>
Co-authored-by: RichardLussier <lussi036@morris.umn.edu>