Add Start Condition to Prevent the Game Dividing By 0 or Ending at 1
This commit adds a start condition which checks to see if the target size for the player's stack is less than 2. If that is true, the game will set the target size to 2 to avoid the issue of dividing by zero or ending the game with 1 character (the player) in the stack.