Create Ghost Wall and Make it Spooky when the Stack has the GhostAlly
This commit creates the GhostWall prefab. It updates the StackController so that when the stack contains the GhostAlly, the GhostWalls (or floors) do not collide with the player or the allies. When all GhostAllies in the stack are removed, the player can once again collide with the walls. 

Also updates the ally prefabs so that everything displays properly when passing through the GhostWalls.

Co-authored-by:  ErikRauer <rauer007@morris.umn.edu>