Organize level 1 and modify initial zoom
Organized the prefabs in level 1 into several folders so the hierarchy looks less messy. Additionally, changed the orthographic size of the camera from 5 to 9 to start the player more zoomed out so they can see more. 

Added various text tips to level 1 and modified the level to be meant more for beginners. Finally, changed the method of winning from winning with a certain stack size to getting the certain stack size and standing on a "winning platform"

Co-authored-by: day00090 <day00090@morris.umn.edu>