Update Level 1 Graphic.png to Reflect Level 1's Redesign
Level 1's display graphic in the level selector has been updated to reflect level 1's redesign. The magma has been replaced with the evil slime, and the bottom section has been updated to indicate to the player that Benzo must jump on a platform with the right amount of allies to win.