Fix jump reset glitch
There was a glitch where if you landed on the corner of a platform your jumps wouldn't reset. This is now fixed by checking for collisions against enemies and floors separately. This commit also attempts to fix a small issue where you get stuck if you walk really slowly near the edge of where the evil slime spawns in level 1.

Co-authored-by: day00090 <day00090@morris.umn.edu>