Make levels easier for players
Based on playtester feedback, we made it easier to land on enemies successfully. Additionally, we added more leeway to pass levels such as adding an additional slime enemy in level 1 and level 2 in case the player loses one of their allies.

This commit also fixes the bug where you could reset your jumps while being off of the platform.

Co-authored-by: day00090 <day00090@morris.umn.edu>