Fix Frame Rate Issue for Moving Obstacles When the Game is Built
NOTE: UPDATE ALL MOVING PLATFORM SPEEDS IN LEVELS 3 AND BEYOND OR THEY WILL MOVE VERY SLOWLY

This commit fixes an issue that results in moving platforms moving from side to side slowly when the frame rate is low in the game build. A multiplier of Time.deltaTime has been applied to the moveSpeed to adjust the speed correctly.

Co-authored-by: RichardLussier <lussi036@morris.umn.edu>