Add Functioning Elevator
This commit adds a functional, one-way elevator to level 7. Upon walking onto the elevator platform with an electric ally, the elevator will bring the player up to a new section of the level. If the player does not get the first electric enemy upon arriving at the elevator, they can cross over to the other side to try and get another one.