Update sprite for electric enemy
Changes the pivot point of the sprite for the electric enemy. Previously the pivot was on the bottom left, so every time the enemy flipped directions, the box collider would not flip with the sprite. This lead to the box collider being "in front" of the sprite when the enemy is moving to the right.

Also updates the Electric movement to not require a Rigidbody, since we use the transform for all the movement anyway.