Update credits
Updates the credits with the new art that has been added in the last couple of levels. Also makes the credit scene scrollable.

I'm not sure how to limit how far up/down the user can scroll, which we should probably do so that they do not end up with nothing on the screen.