Remove Top Box Collider From Snake
Erik discovered why the snake is so hard to acquire at times: the top box collider is not flipping with the rest of the snake when it changes direction. In order to prevent the issue from hindering player progress during tomorrow's playtest, Erik and I have elected to remove the top box collider entirely.

Co-authored-by: ErikRauer <rauer007@morris.umn.edu>