Move enemy over to the left
There was a skeleton enemy that was too far to the right and sometimes would be on the other side of a mushroom wall, making the level unbeatable. This commit moves the skeleton to the left and decreases its side-to-side distance so it won't be on the other side of the wall.