Update Starting Zoom
This commit updates the camera controller so that the starting zoom on the player is a little further out. Each level has been updated so that the min y of the camera matches with the new level of zoom.

This commit also updates the tips in Level1 so that they are a little less transparent so the player can read them better.